+++ include/mercator_util.h
@@ -14,14 +14,14 @@ namespace mercator_util {
 
    // Computes the X tile cooredinate of quadtree at level Z
    inline uint32_t lon2tilex(double lon, int z) { // z = 25
-      uint32_t x = (lon + 180.0) / 360.0 * (1 << z);
+      uint32_t x = (uint32_t)((lon + 180.0) / 360.0 * (1 << z));
       return x & ((1 << z) - 1);
    }
 
    // Computes the Y tile cooredinate of quadtree at level Z
    inline uint32_t lat2tiley(double lat, int z) {
       static const double PI_180 = M_PI / 180.0;
-      uint32_t y = (1.0 - log(tan(lat * PI_180) + 1.0 / cos(lat * PI_180)) / M_PI) / 2.0 * (1 << z);
+      uint32_t y = (uint32_t)((1.0 - log(tan(lat * PI_180) + 1.0 / cos(lat * PI_180)) / M_PI) / 2.0 * (1 << z));
       return y & ((1 << z) - 1);
    }
